Change in the structure of project (data validation)
